fix: corrige tipo do campo lat de Double para BigDecimal Schema validation do Hibernate falhava no Render/Aiven porque a coluna lat no banco e DECIMAL(12,6) mas a entidade declarava Double (mapeado como float pelo Hibernate). - CandidatoEntity: lat Double -> BigDecimal - CandidatoMatchDTO: lat Double -> BigDecimal, lon mantido como String - MatchingServiceUnitTest: literais double -> new BigDecimal(...) - MatchingJavaPythonParityExportTest: idem 28 testes passando.

+++ backend/src/main/java/br/com/appbit/appbit/entities/CandidatoEntity.java
@@ -10,6 +10,7 @@ import lombok.NoArgsConstructor;
 import lombok.Setter;
 import lombok.ToString;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -52,7 +53,7 @@ public class CandidatoEntity {
     private String lon;
 
     @Column(name = "lat")
-    private Double lat;
+    private BigDecimal lat;
 
     @Column(name = "municipio_residencia")
     private String municipio;
